% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpsurvey-package.R
\docType{package}
\name{ahpsurvey-package}
\alias{ahpsurvey-package}
\title{ahpsurvey: Analytic Hierarchy Process for Survey Data}
\description{
The Analytic Hierarchy Process is a versatile multi-criteria decision-making tool introduced by Saaty (1987) \doi{10.1016/0270-0255(87)90473-8} that allows decision-makers to weigh attributes and evaluate alternatives presented to them. This package provides a consistent methodology for researchers to reformat data and run analytic hierarchy process in R on data that are formatted using the survey data entry mode. It is optimized for performing the analytic hierarchy process with many decision-makers, and provides tools and options for researchers to aggregate individual preferences and test multiple options. It also allows researchers to quantify, visualize and correct for inconsistency in the decision-maker's comparisons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frankiecho.github.io/ahpsurvey/}
  \item \url{https://github.com/frankiecho/ahpsurvey}
  \item Report bugs at \url{https://github.com/frankiecho/ahpsurvey/issues}
}

}
\author{
\strong{Maintainer}: Frankie Cho \email{frankie.cho@monash.edu} [copyright holder]

Authors:
\itemize{
  \item Frankie Cho \email{frankie.cho@monash.edu} [copyright holder]
}

}
\keyword{internal}
