% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpsurvey-package.R
\name{ahpsurvey}
\alias{ahpsurvey}
\title{\code{ahpsurvey}: Analytic Hierarchy Process for Survey Data}
\description{
The \code{ahpsurvey} package provides a workflow for researchers to reformat data and run analytic hierarchy process on data that are originally formatted using the survey data entry mode. A diversified toolbox is available for researchers to explore ways to impute missing pairwise comparisons or to adjust comparisons based on consistency, and to conveniently customize the methods which weights derived from surveys are extracted and aggregated.
}
\author{
Frankie Cho \email{frankie.cho@monash.edu}
}
\keyword{Analytic}
\keyword{Hierarchy}
\keyword{Process}
