% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city1}
\alias{city1}
\title{Dataframe of one decision-maker based on \insertCite{Saaty2004;textual}{ahpsurvey}}
\format{A data frame with 1 row and 10 variables, which are pairwise comparisons of \code{atts}.}
\source{
\insertCite{Saaty2004;textual}{ahpsurvey}
}
\usage{
data(city1)
}
\description{
A data.frame of one decision-maker with weights the same as \insertCite{Saaty2004;textual}{ahpsurvey}, with the pairwise comparisons of \code{atts}: \code{c('cult', 'fam', 'house', 'jobs', 'trans')}, about a choosing the best city to live in based on five attributes: Culture, Family, House, Jobs, and Transportation. Negative values in the data denote that the attribute on the left is more important than the right, thus if used with \code{ahp.mat}, \code{negconvert} must be set to \code{TRUE}.
}
\keyword{datasets}
